# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Color API
  description: API pour gérer les stocks des couleurs.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Entrez votre token JWT (obtenu via /login)"
...
paths:
...
  /colors/list:
    post:
      tags:
        - Color
      summary: Obtenir la liste de toutes les couleurs avec leur disponibilité de stock.
      description: |
        Récupère toutes les couleurs avec leurs informations de disponibilité.
        Les stocks sont binaires : 0 = pas disponible, 1 = disponible.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                filter:
                  type: string
                  enum: ["shiny_stock", "matte_stock", "sanded_stock", "no_stock"]
                  description: "Filtre par type de disponibilité"
                  example: "shiny_stock"
      responses:
        '200':
          description: Liste des couleurs récupérée avec succès.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    type:
                      type: string
                      example: "RAL"
                    name_fr:
                      type: string
                      example: "Rouge"
                    name_en:
                      type: string
                      example: "Red"
                    name_pt:
                      type: string
                      example: "Vermelho"
                    value:
                      type: string
                      example: "RAL3000"
                    color:
                      type: string
                      example: "#FF0000"
                    shiny_stock:
                      type: integer
                      enum: [0, 1]
                      description: "Disponibilité stock brillant (0 = pas de stock, 1 = en stock)"
                      example: 1
                    matte_stock:
                      type: integer
                      enum: [0, 1]
                      description: "Disponibilité stock mat (0 = pas de stock, 1 = en stock)"
                      example: 0
                    sanded_stock:
                      type: integer
                      enum: [0, 1]
                      description: "Disponibilité stock sablée (0 = pas de stock, 1 = en stock)"
                      example: 1
                    deleted:
                      type: integer
                      example: 0
        '401':
          description: Non autorisé (token manquant ou invalide).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "No token provided"
                  status:
                    type: integer
                    example: 401
        '500':
          description: Erreur serveur.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Server error"
  /colors/modifyStock/{id}:
    post:
      tags:
        - Color
      summary: Modifier le stock d'une couleur spécifique.
      description: |
        ⚠️ **Permissions requises** : admin ou color manager
        
        Modifie la disponibilité des stocks (shiny, matte, sanded) d'une couleur.
        Les valeurs sont binaires : 0 = pas de stock, 1 = en stock.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: "ID unique de la couleur à modifier"
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shiny_stock:
                  type: integer
                  enum: [0, 1]
                  description: "Disponibilité stock brillant (0 = pas de stock, 1 = en stock)"
                  example: 1
                matte_stock:
                  type: integer
                  enum: [0, 1]
                  description: "Disponibilité stock mat (0 = pas de stock, 1 = en stock)"
                  example: 0
                sanded_stock:
                  type: integer
                  enum: [0, 1]
                  description: "Disponibilité stock sablée (0 = pas de stock, 1 = en stock)"
                  example: 1
              required:
                - shiny_stock
                - matte_stock
                - sanded_stock
      responses:
        '200':
          description: Stock modifié avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Color stock updated successfully"
                  status:
                    type: integer
                    example: 200
                  valid:
                    type: boolean
                    example: true
        '400':
          description: Données invalides (ID manquant ou valeurs de stock non valides - doivent être 0 ou 1).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Stock values must be 0 or 1"
                  valid:
                    type: boolean
                    example: false
        '401':
          description: Token manquant ou invalide.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token requis"
        '403':
          description: Permissions insuffisantes (nécessite admin ou color manager).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Permissions insuffisantes"
        '404':
          description: Couleur non trouvée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Color not found"
                  valid:
                    type: boolean
                    example: false
        '500':
          description: Erreur serveur.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server error"
                  status:
                    type: integer
                    example: 500
  /colors/addColor:
    post:
      tags:
        - Admin
      summary: Ajouter une nouvelle couleur.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Ajoute une nouvelle couleur avec sa disponibilité de stock initiale.
        Les stocks sont binaires : 0 = pas disponible, 1 = disponible.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nameFr:
                  type: string
                  description: "Nom de la couleur"
                  example: "Rouge Cardinal"
                nameEn:
                  type: string
                  description: "Nom de la couleur en anglais"
                  example: "Cardinal Red"
                namePt:
                  type: string
                  description: "Nom de la couleur en portugais"
                  example: "Vermelho Cardeal"
                value:
                  type: string
                  description: "Code de la couleur (ex: RAL3000)"
                  example: "RAL3000"
                type:
                  type: string
                  description: "Type de la couleur (ex: RAL, Pantone)"
                  example: "RAL"
                deleted:
                  type: integer
                  description: "Statut de suppression (0 = actif, 1 = supprimé)"
                  example: 0
                color:
                  type: string
                  description: "Code hexadécimal de la couleur"
                  example: "#FF0000"
                shiny_stock:
                  type: integer
                  enum: [0, 1]
                  description: "Disponibilité stock brillant initial (0 = pas de stock, 1 = en stock)"
                  example: 1
                matte_stock:
                  type: integer
                  enum: [0, 1]
                  description: "Disponibilité stock mat initial (0 = pas de stock, 1 = en stock)"
                  example: 0
                sanded_stock:
                  type: integer
                  enum: [0, 1]
                  description: "Disponibilité stock sablée initial (0 = pas de stock, 1 = en stock)"
                  example: 1
              required:
                - nameFr
                - nameEn
                - namePt
                - type
                - value
                - color
                - shiny_stock
                - matte_stock
                - sanded_stock
      responses:
        '201':
          description: Couleur créée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Color added successfully"
                  status:
                    type: integer
                    example: 201
                  valid:
                    type: boolean
                    example: true
                  value:
                    type: object
                    description: "Données de la couleur créée"
        '400':
          description: Données invalides ou manquantes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid data provided"
                  valid:
                    type: boolean
                    example: false
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '500':
          description: Erreur serveur.
  /colors/deleteColor/{id}:
    post:
      tags:
        - Admin
      summary: Supprimer une couleur (suppression logique).
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Effectue une suppression logique de la couleur (marque comme supprimée).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: "ID unique de la couleur à supprimer"
          example: 1
      responses:
        '200':
          description: Couleur supprimée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Color deleted successfully"
                  valid:
                    type: boolean
                    example: true
        '400':
          description: ID invalide (doit être un entier positif).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "id must be a positive integer"
                  valid:
                    type: boolean
                    example: false
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '404':
          description: Couleur non trouvée.
        '500':
          description: Erreur serveur.
...